==> R/5_HIA.R <==
#-----------------------------------------------------------------------------#
#         5. Population weighted and spatial exposures, RR and PAFs           #
#-----------------------------------------------------------------------------#


# pathroot <- ""
pathroot <- "/PROJECTES/AIRPOLLUTION/lara/LCDEUrban2026_wildfires/"



# 1. RR by Anna Alari ----

# lag0: 1.008 (95% CI 1.006-1.010) per 1ug/m3 increment increase
set.seed(1234)
point_RR10 <- 1.008^10
nsim <- 200
sim_RR10 <- exp(rnorm(nsim, mean = log(point_RR10), sd = log((1.010^10)/(1.006^10))/(qnorm(0.975)*2)))
names(sim_RR10) <- paste0("RR10_", str_pad(1:nsim, 3, "left", "0"))



# 2. Mortality ----

mortality <- fread(paste0(pathroot, "data/processed/mortality.csv"))
mortality <- mortality[, year := as.numeric(year)][year > 2002 & year < 2024][, .(URAU_CODE, date, deaths_urb_merge_day)]

#####
# # mortality for sensitivity analyses
# mortality <- mortality %>% filter(!is.na(deaths_urb_season_day))
# # Don't forget to change the deaths variable in the function (ie. deaths_urb_season_day or deaths_total_day
# # for deaths with seasonality only or deaths with no seasonality, respectively)
#####



# 3. HIA Function ----

hia <- function(expofile, pRR10 = point_RR10, sRR10 = sim_RR10, mort = mortality){

  # Read exposure data, delete cells with no population
  expodata <- setDT(fread(expofile))
  expodata <- expodata[, .(URAU_CODE, citysize, CNTR_CODE, region, pop, date, pm25)]
  expodata[, c("URAU_CODE", "citysize", "CNTR_CODE", "region") := lapply(.SD, as.character), 
           .SDcols = c("URAU_CODE", "citysize", "CNTR_CODE", "region")]
  
  # Checks
  if(any(tapply(expodata$pop, expodata$URAU_CODE, sum) == 0)){
    stop("Some cities have a population of 0. Please revise the data.")
  }
  if(any(!mort$URAU_CODE %in% expodata$URAU_CODE)){
    stop("Missing exposure data for some cities.")
  }
  
  # Compute RRs
  expodata$RR <- exp(log(pRR10)*expodata$pm25/10)
  expodata_sims <- map_dfc(sRR10, function(x) exp(log(x)*expodata$pm25/10))
  names(expodata_sims) <- gsub("RR10", "RR", names(expodata_sims))
  expodata <- cbind(expodata, expodata_sims)
  rm("expodata_sims")
  
  # Compute population weights by city, daily population weighted exposure + PAFs
  setDT(expodata)
  expodata[, popw := pop/sum(pop), by = .(URAU_CODE, citysize, CNTR_CODE, region, date)]
  RRcols <- names(expodata)[grepl("RR", names(expodata))]
  expodata <- expodata[, c(pm25w=sum(popw*pm25),
                           lapply(.SD, function(x) (sum(popw*x)-1)/sum(popw*x))),
                       by = .(URAU_CODE, citysize, CNTR_CODE, region, date), .SDcols = RRcols]
  names(expodata) <- gsub("RR", "PAF", names(expodata))
  
  # Merge mortality and exposure, and estimate attributable mortality
  setkey(mort, URAU_CODE, date)
  setkey(expodata, URAU_CODE, date)
  mort_city <- mort[expodata, nomatch = 0]
  PAFcols <- names(mort_city)[grepl("PAF", names(mort_city))]
  # for sensitivity change deaths variable to either deaths_urb_season_day or deaths_total_day
  mort_city <- mort_city[, lapply(.SD, function(x) x * deaths_urb_merge_day), 
                         by = .(URAU_CODE, citysize, CNTR_CODE, region, date), .SDcols = PAFcols]
  names(mort_city) <- gsub("PAF", "attr", names(mort_city))
  
  # Aggregate to years 
  attrcols <- names(mort_city)[grepl("attr", names(mort_city))]
  mort_city[, year := year(date)]
  mort_city_year <- mort_city[, c(N = .N, lapply(.SD, sum)),
                              by = .(URAU_CODE, citysize, CNTR_CODE, region, year),
                              .SDcols = attrcols]
  
  # Aggregate to city size
  mort_size_year <- mort_city_year %>%
    group_by(citysize, year) %>%
    summarise(Ncities = n(),
              across(starts_with("attr"), sum),
              .groups = "drop")
  
  # Aggregate to country 
  mort_country_year <- mort_city_year %>%
    group_by(year, CNTR_CODE, region) %>%
    summarise(Ncities = n(),
              across(starts_with("attr"), sum),
              .groups = "drop")
  
  # Aggregate to regions 
  mort_region_year <- mort_country_year %>%
    group_by(year, region) %>%
    summarise(Ncountries = n(),
              across(starts_with("attr"), sum),
              .groups = "drop") 
  
  # Aggregate to Europe
  mort_euro_year <- mort_country_year %>%
    group_by(year) %>%
    summarise(Ncountries = n(),
              across(starts_with("attr"), sum),
              .groups = "drop") 
  
  # Compute CIs
  mort_city_year$attrlower <-
    apply(dplyr::select(mort_city_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.025, na.rm=T))
  mort_city_year$attrupper <-
    apply(dplyr::select(mort_city_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.975, na.rm=T))
  mort_size_year$attrlower <-
    apply(dplyr::select(mort_size_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.025, na.rm=T))
  mort_size_year$attrupper <-
    apply(dplyr::select(mort_size_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.975, na.rm=T))
  mort_country_year$attrlower <-
    apply(dplyr::select(mort_country_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.025, na.rm=T))
  mort_country_year$attrupper <-
    apply(dplyr::select(mort_country_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.975, na.rm=T))
  mort_region_year$attrlower <-
    apply(dplyr::select(mort_region_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.025, na.rm=T))
  mort_region_year$attrupper <-
    apply(dplyr::select(mort_region_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.975, na.rm=T))
  mort_euro_year$attrlower <-
    apply(dplyr::select(mort_euro_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.025, na.rm=T))
  mort_euro_year$attrupper <-
    apply(dplyr::select(mort_euro_year, starts_with("attr_")), 1,
          function(x) quantile(x, 0.975, na.rm=T))
  mort_city_year <- dplyr::select(mort_city_year, -starts_with("attr_"))
  mort_size_year <- dplyr::select(mort_size_year, -starts_with("attr_"))
  mort_country_year <- dplyr::select(mort_country_year, -starts_with("attr_"))
  mort_region_year <- dplyr::select(mort_region_year, -starts_with("attr_"))
  mort_euro_year <- dplyr::select(mort_euro_year, -starts_with("attr_"))
  
  # Return data
  list("mortality_city" = mort_city_year,
       "mortality_citysize" = mort_size_year,
       "mortality_country" = mort_country_year,
       "mortality_region" = mort_region_year,
       "mortality_euro" = mort_euro_year)
}



# 4. Execute  ----

hia03 <- hia(paste0(pathroot, "data/processed/assembled/data_2003.csv"))
hia04 <- hia(paste0(pathroot, "data/processed/assembled/data_2004.csv"))
hia05 <- hia(paste0(pathroot, "data/processed/assembled/data_2005.csv"))
hia06 <- hia(paste0(pathroot, "data/processed/assembled/data_2006.csv"))
hia07 <- hia(paste0(pathroot, "data/processed/assembled/data_2007.csv"))
hia08 <- hia(paste0(pathroot, "data/processed/assembled/data_2008.csv"))
hia09 <- hia(paste0(pathroot, "data/processed/assembled/data_2009.csv"))
hia10 <- hia(paste0(pathroot, "data/processed/assembled/data_2010.csv"))
hia11 <- hia(paste0(pathroot, "data/processed/assembled/data_2011.csv"))
hia12 <- hia(paste0(pathroot, "data/processed/assembled/data_2012.csv"))
hia13 <- hia(paste0(pathroot, "data/processed/assembled/data_2013.csv"))
hia14 <- hia(paste0(pathroot, "data/processed/assembled/data_2014.csv"))
hia15 <- hia(paste0(pathroot, "data/processed/assembled/data_2015.csv"))
hia16 <- hia(paste0(pathroot, "data/processed/assembled/data_2016.csv"))
hia17 <- hia(paste0(pathroot, "data/processed/assembled/data_2017.csv"))
hia18 <- hia(paste0(pathroot, "data/processed/assembled/data_2018.csv"))
hia19 <- hia(paste0(pathroot, "data/processed/assembled/data_2019.csv"))
hia20 <- hia(paste0(pathroot, "data/processed/assembled/data_2020.csv"))
hia21 <- hia(paste0(pathroot, "data/processed/assembled/data_2021.csv"))
hia22 <- hia(paste0(pathroot, "data/processed/assembled/data_2022.csv"))
hia23 <- hia(paste0(pathroot, "data/processed/assembled/data_2023.csv"))


# Attributable mortality by city
attr_city <- bind_rows(hia03[[1]], hia04[[1]], hia05[[1]], hia06[[1]], hia07[[1]],
                       hia08[[1]], hia09[[1]], hia10[[1]], hia11[[1]], hia12[[1]],
                       hia13[[1]], hia14[[1]], hia15[[1]], hia16[[1]], hia17[[1]],
                       hia18[[1]], hia19[[1]], hia20[[1]], hia21[[1]], hia22[[1]], 
                       hia23[[1]])
write_csv(attr_city, paste0(pathroot, "data/processed/attributable_city.csv"))

# Attributable mortality by city size
attr_citysize <- bind_rows(hia03[[2]], hia04[[2]], hia05[[2]], hia06[[2]], hia07[[2]],
                          hia08[[2]], hia09[[2]], hia10[[2]], hia11[[2]], hia12[[2]],
                          hia13[[2]], hia14[[2]], hia15[[2]], hia16[[2]], hia17[[2]],
                          hia18[[2]], hia19[[2]], hia20[[2]], hia21[[2]], hia22[[2]], 
                          hia23[[2]])
write_csv(attr_citysize, paste0(pathroot, "data/processed/attributable_citysize.csv"))

# Attributable mortality by country
attr_country <- bind_rows(hia03[[3]], hia04[[3]], hia05[[3]], hia06[[3]], hia07[[3]],
                                 hia08[[3]], hia09[[3]], hia10[[3]], hia11[[3]], hia12[[3]],
                                 hia13[[3]], hia14[[3]], hia15[[3]], hia16[[3]], hia17[[3]],
                                 hia18[[3]], hia19[[3]], hia20[[3]], hia21[[3]], hia22[[3]], 
                                 hia23[[3]])
write_csv(attr_country, paste0(pathroot, "data/processed/attributable_country.csv"))

# Attributable mortality by European region
attr_region <- bind_rows(hia03[[4]], hia04[[4]], hia05[[4]], hia06[[4]], hia07[[4]],
                     hia08[[4]], hia09[[4]], hia10[[4]], hia11[[4]], hia12[[4]],
                     hia13[[4]], hia14[[4]], hia15[[4]], hia16[[4]], hia17[[4]],
                     hia18[[4]], hia19[[4]], hia20[[4]], hia21[[4]], hia22[[4]], 
                     hia23[[4]])
write_csv(attr_region, paste0(pathroot, "data/processed/attributable_region.csv"))

# Attributable mortality euro
attr_euro <- bind_rows(hia03[[5]], hia04[[5]], hia05[[5]], hia06[[5]], hia07[[5]],
                       hia08[[5]], hia09[[5]], hia10[[5]], hia11[[5]], hia12[[5]],
                       hia13[[5]], hia14[[5]], hia15[[5]], hia16[[5]], hia17[[5]],
                       hia18[[5]], hia19[[5]], hia20[[5]], hia21[[5]], hia22[[5]],
                       hia23[[5]])
write_csv(attr_euro, paste0(pathroot, "data/processed/attributable_euro.csv"))

# clean
rm(list = ls())
